refactor: organize features structure
