+++ src/features/dashboard/DeviceCard.jsx
