+++ src/features/dashboard/DeviceCard.module.css
