+++ src/features/dashboard/SensorDashboard.module.css
