+++ src/features/dashboard/SystemTabs.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from '../SensorDashboard.module.css';
+import styles from './SensorDashboard.module.css';
 
 function SystemTabs({systems = [], activeSystem, onChange}) {
     return (
