+++ src/features/dashboard/pages/DashboardPage.jsx
@@ -1,9 +1,9 @@
 // src/pages/DashboardPage.jsx
 /* eslint-disable react-refresh/only-export-components */
 import React, {useEffect, useMemo, useState} from "react";
-import {useLiveNow} from "../hooks/useLiveNow";
-import {SystemOverviewCard, LayerPanel} from "./SystemAndLayerCards";
-import FilterBar from "../components/dashboard/FilterBar";
+import { useLiveNow } from "../useLiveNow";
+import { SystemOverviewCard, LayerPanel } from "./SystemAndLayerCards";
+import FilterBar from "../FilterBar";
 
 // If you don't have a CSS module for this page, keep styles as empty.
 const styles = {};
