+++ src/features/dashboard/pages/Live.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SensorDashboard from '../components/SensorDashboard';
+import SensorDashboard from '../SensorDashboard';
 
 function Live() {
     return <SensorDashboard view="live" title="Live" />;
