+++ src/features/reports/AutoRefreshControls.jsx
